Constructed UpdateProductService, fixed error with adding new product

+++ Frontend/PriceTag/app/src/main/java/antunmod/projects/pricetag/transfer/AddStoreSpecificProductStore.java
@@ -15,12 +15,12 @@ import java.io.Serializable;
 public class AddStoreSpecificProductStore implements Serializable{
 
 	private Short productSpecificId;
-	private byte storeId;
+	private Short storeId;
 	private String storeAddress;
 	private Short userId;
 	private Float price;
 
-	public AddStoreSpecificProductStore(byte storeId, String storeAddress, Short userId, Float price) {
+	public AddStoreSpecificProductStore(Short storeId, String storeAddress, Short userId, Float price) {
 		super();
 		this.storeId = storeId;
 		this.storeAddress = storeAddress;
@@ -28,7 +28,7 @@ public class AddStoreSpecificProductStore implements Serializable{
 		this.price = price;
 	}
 
-	public byte getStoreId() {
+	public Short getStoreId() {
 		return storeId;
 	}
